experimental windows io support

+++ include/util/FileIO.hpp
@@ -1,16 +1,17 @@
 #pragma once
 #include <cstddef>
 #include <string>
+#include <cstdint>
 
 namespace util {
 namespace file {
-    using FileHandle = int;
+    using FileHandle = intptr_t;
     
     FileHandle open_read(const std::string& path);
     FileHandle open_write(const std::string& path);
     void close(FileHandle fd);
     
-    size_t pread(FileHandle fd, void* buf, size_t count, size_t offset);
-    size_t pwrite(FileHandle fd, const void* buf, size_t count, size_t offset);
+    int pread(FileHandle fd, void* buf, size_t count, size_t offset);
+    int pwrite(FileHandle fd, const void* buf, size_t count, size_t offset);
 }
 }
